load_word2vec_format returns KeyedVector, minor refactoring
